Move filter expression code to filter.py
